Add window manager customization
This patch sets up sloppy focus on mouseover. That means that if
a user mouses over a window, that window will be given focus.
HOWEVER if a user then alt-tabs, the alt-tabbed window will be
given focus as expected.

This patch also enables raise-on-click behaviour, which brings a
clicked window to the front of the page. That's behaviour that I
generally want and find useful.